Extract pure core from groove_utils into focused ES modules
Step 2 of the refactor: pull the dependency-free logic out of the ~3400-line
GrooveUtils god-object into their own modules. Strangler-fig — GrooveUtils keeps
thin delegating methods, so callers and the existing suites are unchanged.

- js/constants.js: the 62 shared drum-notation / MIDI constants
- js/urlSerialization.js: groove <-> URL (getGrooveDataFromUrlString / getUrlStringFromGrooveData)
- js/midiFile.js: grooveData -> MIDI file (create_MIDIURLFromGrooveData + helpers)
- js/abcNotation.js: grooveData -> ABC notation (createABCFromGrooveData + the ABC helper cluster)

The extracted functions take a GrooveUtils instance for the helpers that remain
in GrooveUtils and import their constants from constants.js. groove_utils.js
shrinks ~3400 -> 2410 lines.

Tests: existing url/midi/abc suites pass via delegation; added direct-import
tests asserting the modules produce byte-identical output to the delegating
methods. ESLint sourceType=module for the new files.

697 unit + 71 e2e green; golden-master SVG/MIDI snapshots unchanged.
