+++ js/groove_utils.js
@@ -27,6 +27,81 @@
 // eslint.config.js (they're shared across multiple js/ files there, unlike
 // this file's own GrooveUtils global below).
 
+// Shared constants (moved to their own module in Step 2).
+import {
+  constant_DEFAULT_TEMPO,
+  constant_ABC_STICK_R,
+  constant_ABC_STICK_L,
+  constant_ABC_STICK_BOTH,
+  constant_ABC_STICK_COUNT,
+  constant_ABC_STICK_OFF,
+  constant_ABC_HH_Ride,
+  constant_ABC_HH_Ride_Bell,
+  constant_ABC_HH_Cow_Bell,
+  constant_ABC_HH_Crash,
+  constant_ABC_HH_Stacker,
+  constant_ABC_HH_Metronome_Normal,
+  constant_ABC_HH_Metronome_Accent,
+  constant_ABC_HH_Open,
+  constant_ABC_HH_Close,
+  constant_ABC_HH_Accent,
+  constant_ABC_HH_Normal,
+  constant_ABC_SN_Ghost,
+  constant_ABC_SN_Accent,
+  constant_ABC_SN_Normal,
+  constant_ABC_SN_XStick,
+  constant_ABC_SN_Buzz,
+  constant_ABC_SN_Flam,
+  constant_ABC_SN_Drag,
+  constant_ABC_KI_SandK,
+  constant_ABC_KI_Splash,
+  constant_ABC_KI_Normal,
+  constant_ABC_T1_Normal,
+  constant_ABC_T2_Normal,
+  constant_ABC_T3_Normal,
+  constant_ABC_T4_Normal,
+  constant_NUMBER_OF_TOMS,
+  constant_ABC_OFF,
+  constant_OUR_MIDI_METRONOME_1,
+  constant_OUR_MIDI_METRONOME_NORMAL,
+  constant_OUR_MIDI_HIHAT_NORMAL,
+  constant_OUR_MIDI_HIHAT_OPEN,
+  constant_OUR_MIDI_HIHAT_ACCENT,
+  constant_OUR_MIDI_HIHAT_CRASH,
+  constant_OUR_MIDI_HIHAT_STACKER,
+  constant_OUR_MIDI_HIHAT_METRONOME_NORMAL,
+  constant_OUR_MIDI_HIHAT_RIDE,
+  constant_OUR_MIDI_HIHAT_RIDE_BELL,
+  constant_OUR_MIDI_HIHAT_COW_BELL,
+  constant_OUR_MIDI_HIHAT_FOOT,
+  constant_OUR_MIDI_SNARE_NORMAL,
+  constant_OUR_MIDI_SNARE_ACCENT,
+  constant_OUR_MIDI_SNARE_GHOST,
+  constant_OUR_MIDI_SNARE_XSTICK,
+  constant_OUR_MIDI_SNARE_BUZZ,
+  constant_OUR_MIDI_SNARE_FLAM,
+  constant_OUR_MIDI_SNARE_DRAG,
+  constant_OUR_MIDI_KICK_NORMAL,
+  constant_OUR_MIDI_TOM1_NORMAL,
+  constant_OUR_MIDI_TOM2_NORMAL,
+  constant_OUR_MIDI_TOM3_NORMAL,
+  constant_OUR_MIDI_TOM4_NORMAL,
+} from './constants.js';
+import {
+  getGrooveDataFromUrlString as _urlParse,
+  getUrlStringFromGrooveData as _urlBuild,
+} from './urlSerialization.js';
+import {
+  MIDI_build_midi_url_count_in_track as _MIDI_build_midi_url_count_in_track,
+  MIDI_from_HH_Snare_Kick_Arrays as _MIDI_from_HH_Snare_Kick_Arrays,
+  create_MIDIURLFromGrooveData as _create_MIDIURLFromGrooveData,
+} from './midiFile.js';
+import {
+  get_top_ABC_BoilerPlate as _get_top_ABC_BoilerPlate,
+  create_ABC_from_snare_HH_kick_arrays as _create_ABC_from_snare_HH_kick_arrays,
+  createABCFromGrooveData as _createABCFromGrooveData,
+} from './abcNotation.js';
+
 var global_num_GrooveUtilsCreated = 0;
 // This module's own URL, used by getGrooveUtilsBaseLocation() to locate sibling
 // assets (soundfont/, images/). As an ES module, document.currentScript is null,
@@ -35,69 +110,6 @@ var global_grooveUtilsScriptSrc = import.meta.url;
 var global_midiInitialized = false;
 
 // global constants
-var constant_MAX_MEASURES = 10;
-var constant_DEFAULT_TEMPO = 80;
-var constant_ABC_STICK_R = '"R"x';
-var constant_ABC_STICK_L = '"L"x';
-var constant_ABC_STICK_BOTH = '"R/L"x';
-var constant_ABC_STICK_COUNT = '"count"x';
-var constant_ABC_STICK_OFF = '""x';
-var constant_ABC_HH_Ride = "^A'";
-var constant_ABC_HH_Ride_Bell = "^B'";
-var constant_ABC_HH_Cow_Bell = "^D'";
-var constant_ABC_HH_Crash = "^c'";
-var constant_ABC_HH_Stacker = "^d'";
-var constant_ABC_HH_Metronome_Normal = "^e'";
-var constant_ABC_HH_Metronome_Accent = "^f'";
-var constant_ABC_HH_Open = '!open!^g';
-var constant_ABC_HH_Close = '!plus!^g';
-var constant_ABC_HH_Accent = '!accent!^g';
-var constant_ABC_HH_Normal = '^g';
-var constant_ABC_SN_Ghost = '!(.!!).!c';
-var constant_ABC_SN_Accent = '!accent!c';
-var constant_ABC_SN_Normal = 'c';
-var constant_ABC_SN_XStick = '^c';
-var constant_ABC_SN_Buzz = '!///!c';
-var constant_ABC_SN_Flam = '!accent!{/c}c';
-var constant_ABC_SN_Drag = '{/cc}c';
-var constant_ABC_KI_SandK = '[F^d,]'; // kick & splash
-var constant_ABC_KI_Splash = '^d,'; // splash only
-var constant_ABC_KI_Normal = 'F';
-var constant_ABC_T1_Normal = 'e';
-var constant_ABC_T2_Normal = 'd';
-var constant_ABC_T3_Normal = 'B';
-var constant_ABC_T4_Normal = 'A';
-var constant_NUMBER_OF_TOMS = 4;
-var constant_ABC_OFF = false;
-
-var constant_OUR_MIDI_VELOCITY_NORMAL = 85;
-var constant_OUR_MIDI_VELOCITY_ACCENT = 120;
-var constant_OUR_MIDI_VELOCITY_GHOST = 50;
-var constant_OUR_MIDI_METRONOME_1 = 76;
-var constant_OUR_MIDI_METRONOME_NORMAL = 77;
-var constant_OUR_MIDI_HIHAT_NORMAL = 42;
-var constant_OUR_MIDI_HIHAT_OPEN = 46;
-var constant_OUR_MIDI_HIHAT_ACCENT = 108;
-var constant_OUR_MIDI_HIHAT_CRASH = 49;
-var constant_OUR_MIDI_HIHAT_STACKER = 52;
-var constant_OUR_MIDI_HIHAT_METRONOME_NORMAL = 77;
-var constant_OUR_MIDI_HIHAT_METRONOME_ACCENT = 76;
-var constant_OUR_MIDI_HIHAT_RIDE = 51;
-var constant_OUR_MIDI_HIHAT_RIDE_BELL = 53;
-var constant_OUR_MIDI_HIHAT_COW_BELL = 105;
-var constant_OUR_MIDI_HIHAT_FOOT = 44;
-var constant_OUR_MIDI_SNARE_NORMAL = 38;
-var constant_OUR_MIDI_SNARE_ACCENT = 22;
-var constant_OUR_MIDI_SNARE_GHOST = 21;
-var constant_OUR_MIDI_SNARE_XSTICK = 37;
-var constant_OUR_MIDI_SNARE_BUZZ = 104;
-var constant_OUR_MIDI_SNARE_FLAM = 107;
-var constant_OUR_MIDI_SNARE_DRAG = 103;
-var constant_OUR_MIDI_KICK_NORMAL = 35;
-var constant_OUR_MIDI_TOM1_NORMAL = 48;
-var constant_OUR_MIDI_TOM2_NORMAL = 47;
-var constant_OUR_MIDI_TOM3_NORMAL = 45;
-var constant_OUR_MIDI_TOM4_NORMAL = 43;
 
 // make these global so that they are shared among all the GrooveUtils classes invoked
 var global_current_midi_start_time = 0;
@@ -954,177 +966,7 @@ function GrooveUtils() {
   };
 
   root.getGrooveDataFromUrlString = function (encodedURLData) {
-    var Stickings_string;
-    var HH_string;
-    var Snare_string;
-    var Kick_string;
-    var myGrooveData = new root.grooveDataNew();
-    var i;
-
-    myGrooveData.debugMode = parseInt(
-      root.getQueryVariableFromString('Debug', root.debugMode, encodedURLData),
-      10
-    );
-
-    var timeSigArray = root.parseTimeSigString(
-      root.getQueryVariableFromString('TimeSig', '4/4', encodedURLData)
-    );
-    myGrooveData.numBeats = timeSigArray[0];
-    myGrooveData.noteValue = timeSigArray[1];
-
-    myGrooveData.timeDivision = parseInt(
-      root.getQueryVariableFromString('Div', 16, encodedURLData),
-      10
-    );
-    myGrooveData.notesPerMeasure = root.calc_notes_per_measure(
-      myGrooveData.timeDivision,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-
-    myGrooveData.metronomeFrequency = parseInt(
-      root.getQueryVariableFromString('MetronomeFreq', '0', encodedURLData),
-      10
-    );
-
-    myGrooveData.numberOfMeasures = parseInt(
-      root.getQueryVariableFromString('measures', 1, encodedURLData),
-      10
-    );
-    if (myGrooveData.numberOfMeasures < 1 || isNaN(myGrooveData.numberOfMeasures))
-      myGrooveData.numberOfMeasures = 1;
-    else if (myGrooveData.numberOfMeasures > constant_MAX_MEASURES)
-      myGrooveData.numberOfMeasures = constant_MAX_MEASURES;
-
-    Stickings_string = root.getQueryVariableFromString('Stickings', false, encodedURLData);
-    if (!Stickings_string) {
-      Stickings_string = root.GetDefaultStickingsGroove(
-        myGrooveData.notesPerMeasure,
-        myGrooveData.numBeats,
-        myGrooveData.noteValue,
-        myGrooveData.numberOfMeasures
-      );
-      myGrooveData.showStickings = false;
-    } else {
-      myGrooveData.showStickings = true;
-    }
-
-    HH_string = root.getQueryVariableFromString('H', false, encodedURLData);
-    if (!HH_string) {
-      root.getQueryVariableFromString('HH', false, encodedURLData);
-      if (!HH_string) {
-        HH_string = root.GetDefaultHHGroove(
-          myGrooveData.notesPerMeasure,
-          myGrooveData.numBeats,
-          myGrooveData.noteValue,
-          myGrooveData.numberOfMeasures
-        );
-      }
-    }
-
-    Snare_string = root.getQueryVariableFromString('S', false, encodedURLData);
-    if (!Snare_string) {
-      Snare_string = root.GetDefaultSnareGroove(
-        myGrooveData.notesPerMeasure,
-        myGrooveData.numBeats,
-        myGrooveData.noteValue,
-        myGrooveData.numberOfMeasures
-      );
-    }
-
-    Kick_string = root.getQueryVariableFromString('K', false, encodedURLData);
-    if (!Kick_string) {
-      root.getQueryVariableFromString('B', false, encodedURLData);
-      if (!Kick_string) {
-        Kick_string = root.GetDefaultKickGroove(
-          myGrooveData.notesPerMeasure,
-          myGrooveData.numBeats,
-          myGrooveData.noteValue,
-          myGrooveData.numberOfMeasures
-        );
-      }
-    }
-
-    // Get the Toms
-    for (i = 0; i < 4; i++) {
-      // toms are named T1, T2, T3, T4
-      var Tom_string = root.getQueryVariableFromString('T' + (i + 1), false, encodedURLData);
-      if (!Tom_string) {
-        Tom_string = root.GetDefaultTomGroove(
-          myGrooveData.notesPerMeasure,
-          myGrooveData.numBeats,
-          myGrooveData.noteValue,
-          myGrooveData.numberOfMeasures
-        );
-      } else {
-        myGrooveData.showToms = true;
-      }
-
-      /// the toms array index starts at zero (0) the first one is T1
-      myGrooveData.toms_array[i] = root.noteArraysFromURLData(
-        'T' + (i + 1),
-        Tom_string,
-        myGrooveData.notesPerMeasure,
-        myGrooveData.numberOfMeasures
-      );
-    }
-
-    myGrooveData.sticking_array = root.noteArraysFromURLData(
-      'Stickings',
-      Stickings_string,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numberOfMeasures
-    );
-    myGrooveData.hh_array = root.noteArraysFromURLData(
-      'H',
-      HH_string,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numberOfMeasures
-    );
-    myGrooveData.snare_array = root.noteArraysFromURLData(
-      'S',
-      Snare_string,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numberOfMeasures
-    );
-    myGrooveData.kick_array = root.noteArraysFromURLData(
-      'K',
-      Kick_string,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numberOfMeasures
-    );
-
-    myGrooveData.title = root.getQueryVariableFromString('title', '', encodedURLData);
-    myGrooveData.title = decodeURIComponent(myGrooveData.title);
-    myGrooveData.title = myGrooveData.title.replace(/\+/g, ' ');
-
-    myGrooveData.author = root.getQueryVariableFromString('author', '', encodedURLData);
-    myGrooveData.author = decodeURIComponent(myGrooveData.author);
-    myGrooveData.author = myGrooveData.author.replace(/\+/g, ' ');
-
-    myGrooveData.comments = root.getQueryVariableFromString('comments', '', encodedURLData);
-    myGrooveData.comments = decodeURIComponent(myGrooveData.comments);
-    myGrooveData.comments = myGrooveData.comments.replace(/\+/g, ' ');
-
-    myGrooveData.tempo = parseInt(
-      root.getQueryVariableFromString('tempo', constant_DEFAULT_TEMPO, encodedURLData),
-      10
-    );
-    if (isNaN(myGrooveData.tempo) || myGrooveData.tempo < 20 || myGrooveData.tempo > 400)
-      myGrooveData.tempo = constant_DEFAULT_TEMPO;
-
-    myGrooveData.swingPercent = parseInt(
-      root.getQueryVariableFromString('swing', 0, encodedURLData),
-      10
-    );
-    if (
-      isNaN(myGrooveData.swingPercent) ||
-      myGrooveData.swingPercent < 0 ||
-      myGrooveData.swingPercent > 100
-    )
-      myGrooveData.swingPercent = 0;
-
-    return myGrooveData;
+    return _urlParse(root, encodedURLData);
   };
 
   // get a really long URL that encodes all of the notes and the rest of the state of the page.
@@ -1132,130 +974,7 @@ function GrooveUtils() {
   //
 
   root.getUrlStringFromGrooveData = function (myGrooveData, url_destination) {
-    var fullURL = window.location.protocol + '//' + window.location.host + window.location.pathname;
-
-    if (!url_destination) {
-      // then assume it is the groove writer display.  Do nothing
-    } else if (url_destination == 'display') {
-      // asking for the "groove_display" page
-      if (fullURL.includes('index.html'))
-        fullURL = fullURL.replace('index.html', 'GrooveEmbed.html');
-      else if (fullURL.includes('/gscribe'))
-        fullURL = fullURL.replace('/gscribe', '/groove/GrooveEmbed.html');
-      else fullURL += 'GrooveEmbed.html';
-    } else if (url_destination == 'fullGrooveScribe') {
-      // asking for the full GrooveScribe link
-      fullURL = 'https://www.mikeslessons.com/gscribe';
-    }
-
-    fullURL += '?';
-
-    if (myGrooveData.debugMode) fullURL += 'Debug=1&';
-
-    if (myGrooveData.viewMode) fullURL += 'Mode=view&';
-
-    if (myGrooveData.grooveDBAuthoring) fullURL += 'GDB_Author=1&';
-
-    fullURL += 'TimeSig=' + myGrooveData.numBeats + '/' + myGrooveData.noteValue;
-
-    // # of notes
-    fullURL += '&Div=' + myGrooveData.timeDivision;
-
-    if (myGrooveData.title !== '') fullURL += '&Title=' + encodeURIComponent(myGrooveData.title);
-
-    if (myGrooveData.author !== '') fullURL += '&Author=' + encodeURIComponent(myGrooveData.author);
-
-    if (myGrooveData.comments !== '')
-      fullURL += '&Comments=' + encodeURIComponent(myGrooveData.comments);
-
-    fullURL += '&Tempo=' + myGrooveData.tempo;
-
-    if (myGrooveData.swingPercent > 0) fullURL += '&Swing=' + myGrooveData.swingPercent;
-
-    // # of measures
-    fullURL += '&Measures=' + myGrooveData.numberOfMeasures;
-
-    // # metronome setting
-    if (myGrooveData.metronomeFrequency !== 0) {
-      fullURL += '&MetronomeFreq=' + myGrooveData.metronomeFrequency;
-    }
-
-    // notes
-    var total_notes = myGrooveData.notesPerMeasure * myGrooveData.numberOfMeasures;
-    var HH =
-      '&H=|' +
-      root.tabLineFromAbcNoteArray(
-        'H',
-        myGrooveData.hh_array,
-        true,
-        true,
-        total_notes,
-        myGrooveData.notesPerMeasure
-      );
-    var Snare =
-      '&S=|' +
-      root.tabLineFromAbcNoteArray(
-        'S',
-        myGrooveData.snare_array,
-        true,
-        true,
-        total_notes,
-        myGrooveData.notesPerMeasure
-      );
-    var Kick =
-      '&K=|' +
-      root.tabLineFromAbcNoteArray(
-        'K',
-        myGrooveData.kick_array,
-        true,
-        true,
-        total_notes,
-        myGrooveData.notesPerMeasure
-      );
-
-    fullURL += HH + Snare + Kick;
-
-    // only add if we need them.  // they are long and ugly. :)
-    if (myGrooveData.showToms) {
-      var Tom1 =
-        '&T1=|' +
-        root.tabLineFromAbcNoteArray(
-          'T1',
-          myGrooveData.toms_array[0],
-          true,
-          true,
-          total_notes,
-          myGrooveData.notesPerMeasure
-        );
-      var Tom4 =
-        '&T4=|' +
-        root.tabLineFromAbcNoteArray(
-          'T4',
-          myGrooveData.toms_array[3],
-          true,
-          true,
-          total_notes,
-          myGrooveData.notesPerMeasure
-        );
-      fullURL += Tom1 + Tom4;
-    }
-
-    // only add if we need them.  // they are long and ugly. :)
-    if (myGrooveData.showStickings) {
-      var Stickings =
-        '&Stickings=|' +
-        root.tabLineFromAbcNoteArray(
-          'stickings',
-          myGrooveData.sticking_array,
-          true,
-          true,
-          total_notes,
-          myGrooveData.notesPerMeasure
-        );
-      fullURL += Stickings;
-    }
-
-    return fullURL;
+    return _urlBuild(root, myGrooveData, url_destination);
   };
 
   function setupHotKeys() {
@@ -1315,226 +1034,35 @@ function GrooveUtils() {
     timeSigBottom,
     renderWidth
   ) {
-    // boiler plate
-    var fullABC = '%abc\n%%fullsvg _' + root.grooveUtilsUniqueIndex + '\nX:6\n';
-
-    fullABC += 'M:' + timeSigTop + '/' + timeSigBottom + '\n';
-
-    // always add a Title even if it's blank
-    fullABC += 'T: ' + tuneTitle + '\n';
-
-    if (tuneAuthor !== '') {
-      fullABC += 'C: ' + tuneAuthor + '\n';
-      fullABC += '%%musicspace 20px\n'; // add some more space
-    }
-
-    if (renderWidth < 400) renderWidth = 400; // min-width
-    if (renderWidth > 3000) renderWidth = 3000; // max-width
-    // the width of the music is always 25% bigger than what we pass in.   Go figure.
-    renderWidth = Math.floor(renderWidth * 0.75);
-
-    fullABC += 'L:1/' + 32 + '\n'; // 4/4 = 32,  6/8 = 64
-
-    if (isPermutation) fullABC += '%%stretchlast 0\n';
-    else fullABC += '%%stretchlast 1\n';
-
-    fullABC +=
-      '%%flatbeams 1\n' +
-      '%%ornament up\n' +
-      '%%pagewidth ' +
-      renderWidth +
-      'px\n' +
-      '%%leftmargin 0cm\n' +
-      '%%rightmargin 0cm\n' +
-      '%%topspace 10px\n' +
-      '%%titlefont calibri 20\n' +
-      '%%partsfont calibri 16\n' +
-      '%%gchordfont calibri 16\n' +
-      '%%annotationfont calibri 16\n' +
-      '%%infofont calibri 16\n' +
-      '%%textfont calibri 16\n' +
-      '%%deco (. 0 a 5 1 1 "@-8,-3("\n' +
-      '%%deco ). 0 a 5 1 1 "@4,-3)"\n' +
-      '%%beginsvg\n' +
-      ' <defs>\n' +
-      ' <path id="Xhead" d="m-3,-3 l6,6 m0,-6 l-6,6" class="stroke" style="stroke-width:1.2"/>\n' +
-      ' <path id="Trihead" d="m-3,2 l 6,0 l-3,-6 l-3,6 l6,0" class="stroke" style="stroke-width:1.2"/>\n' +
-      ' </defs>\n' +
-      '%%endsvg\n' +
-      '%%map drum ^g heads=Xhead print=g       % Hi-Hat\n' +
-      "%%map drum ^c' heads=Xhead print=c'   % Crash\n" +
-      "%%map drum ^d' heads=Xhead print=d'   % Stacker\n" +
-      "%%map drum ^e' heads=Xhead print=e'   % Metronome click\n" +
-      "%%map drum ^f' heads=Xhead print=f'   % Metronome beep\n" +
-      "%%map drum ^A' heads=Xhead print=A'   % Ride\n" +
-      "%%map drum ^B' heads=Trihead print=A' % Ride Bell\n" +
-      "%%map drum ^D' heads=Trihead print=g   % Cow Bell\n" +
-      '%%map drum ^c heads=Xhead print=c  % Cross Stick\n' +
-      '%%map drum ^d, heads=Xhead print=d,  % Foot Splash\n';
-
-    //if(kick_stems_up)
-    //fullABC += "%%staves (Stickings Hands)\n";
-    //else
-    fullABC += '%%staves (Stickings Hands Feet)\n';
-
-    // print comments below the legend if there is one, otherwise in the header section
-    if (tuneComments !== '') {
-      fullABC += 'P: ' + tuneComments + '\n';
-      fullABC += '%%musicspace 20px\n'; // add some more space
-    }
-
-    // the K ends the header;
-    fullABC += 'K:C clef=perc\n';
-
-    if (showLegend) {
-      fullABC +=
-        'V:Stickings\n' +
-        'x8 x8 x8 x8 x8 x8 x8 x8 ||\n' +
-        'V:Hands stem=up \n' +
-        '%%voicemap drum\n' +
-        '"^Hi-Hat"^g4 "^Open"!open!^g4 ' +
-        '"^Crash"^c\'4 "^Stacker"^d\'4 "^Ride"^A\'4 "^Ride Bell"^B\'4 x2 "^Tom"e4 "^Tom"A4 "^Snare"c4 "^Buzz"!///!c4 "^Cross"^c4 "^Ghost  "!(.!!).!c4 "^Flam"{/c}c4  x10 ||\n' +
-        'V:Feet stem=down \n' +
-        '%%voicemap drum\n' +
-        'x52 "^Kick"F4 "^HH foot"^d,4 x4 ||\n' +
-        'T:\n';
-    }
-
-    // tempo setting
-    //fullABC += "Q: 1/4=" + getTempo() + "\n";
-
-    return fullABC;
+    return _get_top_ABC_BoilerPlate(
+      root,
+      isPermutation,
+      tuneTitle,
+      tuneAuthor,
+      tuneComments,
+      showLegend,
+      isTriplets,
+      kick_stems_up,
+      timeSigTop,
+      timeSigBottom,
+      renderWidth
+    );
   };
 
   // looks for modifiers like !accent! or !plus! and moves them outside of the group abc array.
   // Most modifiers (but not all) will not render correctly if they are inside the abc group.
   // returns a string that should be added to the abc_notation if found.
-  function moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, modifier_to_look_for) {
-    var found_modifier = false;
-    var rindex = abcNoteStrings.notes1.lastIndexOf(modifier_to_look_for);
-    if (rindex > -1) {
-      found_modifier = true;
-      abcNoteStrings.notes1 = abcNoteStrings.notes1.replace(modifier_to_look_for, '');
-    }
-    rindex = abcNoteStrings.notes2.lastIndexOf(modifier_to_look_for);
-    if (rindex > -1) {
-      found_modifier = true;
-      abcNoteStrings.notes2 = abcNoteStrings.notes2.replace(modifier_to_look_for, '');
-    }
-    rindex = abcNoteStrings.notes3.lastIndexOf(modifier_to_look_for);
-    if (rindex > -1) {
-      found_modifier = true;
-      abcNoteStrings.notes3 = abcNoteStrings.notes3.replace(modifier_to_look_for, '');
-    }
-    if (found_modifier) return modifier_to_look_for;
-
-    return ''; // didn't find it so return nothing
-  }
 
   // take an array of arrays and use a for loop to test to see
   // if all of the arrays are equal to the "test_value" for a given "test_index"
   // returns "true" if they are all equal.
   // returns "false" if any one of them fails
-  function testArrayOfArraysForEquality(array_of_arrays, test_index, test_value) {
-    for (var i = 0; i < array_of_arrays.length; i++) {
-      if (
-        array_of_arrays[i][test_index] !== undefined &&
-        array_of_arrays[i][test_index] !== test_value
-      )
-        return false;
-    }
-
-    return true;
-  }
 
   // note1_array:   an array containing "false" or a note character in ABC to designate that is is on
   // note2_array:   an array containing "false" or a note character in ABC to designate that is is on
   // end_of_group:  when to stop looking ahead in the array.  (since we group notes in to beats)
-  function getABCforNote(note_array_of_arrays, start_index, end_of_group, scaler) {
-    var ABC_String = '';
-    var abcNoteStrings = {
-      notes1: '',
-      notes2: '',
-      notes3: '',
-    };
-    var num_notes_on = 0;
-    var nextCount;
-
-    for (var which_array = 0; which_array < note_array_of_arrays.length; which_array++) {
-      if (
-        note_array_of_arrays[which_array][start_index] !== undefined &&
-        note_array_of_arrays[which_array][start_index] !== false
-      ) {
-        // look ahead and see when the next note is
-        // the length of this note is dependant on when the next note lands
-        // for every empty space we increment nextCount, and then make the note that long
-        nextCount = 1;
-        for (var indexA = start_index + 1; indexA < start_index + end_of_group; indexA++) {
-          if (!testArrayOfArraysForEquality(note_array_of_arrays, indexA, false)) {
-            break;
-          } else {
-            nextCount++;
-          }
-        }
-
-        abcNoteStrings.notes1 +=
-          note_array_of_arrays[which_array][start_index] + scaler * nextCount;
-        num_notes_on++;
-      }
-    }
-
-    if (num_notes_on > 1) {
-      // if multiple are on, we need to combine them with []
-      // horrible hack.  Turns out ABC will render the accents wrong unless the are outside the brackets []
-      // look for any accents that are delimited by "!"  (eg !accent!  or !plus!)
-      // move the accents to the front
-      ABC_String += moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '!accent!');
-      // in case there are two accents (on both snare and hi-hat) we remove the second one
-      moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '!accent!');
-      ABC_String += moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '!plus!');
-      ABC_String += moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '!open!');
-      ABC_String += moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '!///!');
-
-      // Look for '[' and ']'.   They are added on to the the kick and splash and could be added to other notes
-      // in the future.   They imply that the notes are on the same beat.   Since we are already putting multiple
-      // notes on the same beat (see code below this line that adds '[' & ']'), we need to remove them or the
-      // resulting ABC will be invalid
-      moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '[');
-      moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, ']');
-
-      // this is the flam notation, it can't be in a sub grouping
-      ABC_String += moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '{/c}');
-      // this is the drag notation, it can't be in a sub grouping
-      ABC_String += moveAccentsOrOtherModifiersOutsideOfGroup(abcNoteStrings, '{/cc}');
-
-      ABC_String +=
-        '[' + abcNoteStrings.notes1 + abcNoteStrings.notes2 + abcNoteStrings.notes3 + ']'; // [^gc]
-    } else {
-      ABC_String += abcNoteStrings.notes1 + abcNoteStrings.notes2 + abcNoteStrings.notes3; // note this could be a noOp if all strings are blank
-    }
-
-    return ABC_String;
-  }
 
   // calculate the rest ABC string
-  function getABCforRest(note_array_of_arrays, start_index, end_of_group, scaler, use_hidden_rest) {
-    var ABC_String = '';
-
-    // count the # of rest
-    if (testArrayOfArraysForEquality(note_array_of_arrays, start_index, false)) {
-      var restCount = 1;
-      for (var indexB = start_index + 1; indexB < start_index + end_of_group; indexB++) {
-        if (!testArrayOfArraysForEquality(note_array_of_arrays, indexB, false)) break;
-        else restCount++;
-      }
-
-      // now output a rest for the duration of the rest count
-      if (use_hidden_rest) ABC_String += 'x' + scaler * restCount;
-      else ABC_String += 'z' + scaler * restCount;
-    }
-
-    return ABC_String;
-  }
 
   // the note grouping size is how groups of notes within a measure group
   // for 8ths and 16th we group with 4
@@ -1578,24 +1106,6 @@ function GrooveUtils() {
   // scale correctly
   // The base array is now 32 notes long to support 32nd notes
   // since we would normally group by 4 we need to group by 8 since we are scaling it
-  function abc_gen_note_grouping_size(usingTriplets, timeSigTop, timeSigBottom) {
-    var note_grouping;
-
-    if (usingTriplets) {
-      note_grouping = 12;
-    } else if (timeSigTop == 3) {
-      // 3/4, 3/8, 3/16
-      note_grouping = 8 * (4 / timeSigBottom);
-    } else if (timeSigTop % 6 == 0 && timeSigBottom % 8 == 0) {
-      // 3/4, 6/8, 9/8, 12/8
-      note_grouping = 12 * (8 / timeSigBottom);
-    } else {
-      //note_grouping = 8 * (4/timeSigBottom);
-      note_grouping = 8;
-    }
-
-    return note_grouping;
-  }
 
   root.notesPerMeasureInFullSizeArray = function (is_triplet_division, timeSigTop, timeSigBottom) {
     // a full measure will be defined as 8 * timeSigTop.   (4 = 32, 5 = 40, 6 = 48, etc.)
@@ -1659,20 +1169,6 @@ function GrooveUtils() {
 
   // count the number of note positions that are not rests in all the arrays
   // FFFxFFFxF  would be 2
-  function count_active_notes_in_arrays(array_of_arrays, start_index, how_far_to_measure) {
-    var num_active_notes = 0;
-
-    for (var i = start_index; i < start_index + how_far_to_measure; i++) {
-      for (var which_array = 0; which_array < array_of_arrays.length; which_array++) {
-        if (array_of_arrays[which_array][i] !== false) {
-          num_active_notes++;
-          which_array = array_of_arrays.length; // exit this inner for loop immediately
-        }
-      }
-    }
-
-    return num_active_notes;
-  }
 
   // takes 4 arrays 48 elements long that represent the stickings, snare, HH & kick.
   // each element contains either the note value in ABC "F","^g" or false to represent off
@@ -1682,487 +1178,12 @@ function GrooveUtils() {
   // We output 48 notes in the ABC rather than the traditional 16 or 32 for 4/4 time.
   // This is because of the stickings above the bar are a separate voice and should not have the "3" above them
   // This could be changed to using the normal number and moving all the stickings down to be comments on each note in one voice (But is a pretty big change)
-  function snare_HH_kick_ABC_for_triplets(
-    sticking_array,
-    HH_array,
-    snare_array,
-    kick_array,
-    toms_array,
-    post_voice_abc,
-    num_notes,
-    sub_division,
-    notes_per_measure,
-    kick_stems_up,
-    timeSigTop,
-    timeSigBottom,
-    numberOfMeasuresPerLine
-  ) {
-    var scaler = 1; // we are always in 48 notes here, and the ABC needs to think we are in 48 since the specified division is 1/32
-    var ABC_String = '';
-    var stickings_voice_string = 'V:Stickings\n';
-    var hh_snare_voice_string = 'V:Hands stem=up\n%%voicemap drum\n';
-    var kick_voice_string = 'V:Feet stem=down\n%%voicemap drum\n';
-    var all_drum_array_of_array;
-
-    // console.log(HH_array);
-    // console.log(kick_array);
-    // console.log(notes_per_measure);
-    // console.log(sub_division);
-
-    if (kick_stems_up) {
-      all_drum_array_of_array = [snare_array, HH_array, kick_array];
-    } else {
-      all_drum_array_of_array = [snare_array, HH_array]; // exclude the kick
-    }
-    if (toms_array) all_drum_array_of_array = all_drum_array_of_array.concat(toms_array);
-
-    // occationally we will change the sub_division output to 1/8th or 1/16th notes when we detect a beat that is better displayed that way
-    // By default we use the base sub_division but this can be set different below
-    var faker_sub_division = sub_division;
-
-    for (var i = 0; i < num_notes; i++) {
-      // triplets are special.  We want to output a note or a rest for every space of time
-      // 8th note triplets should always use rests
-      // end_of_group should be
-      //  "4" for 1/8th note triplets
-      //  "2" for 1/16th note triplets
-      //  "1" for 1/32nd note triplets.
-      var end_of_group = 48 / faker_sub_division;
-      var grouping_size_for_rests = end_of_group;
-      var skip_adding_more_notes = false;
-
-      if ((i % notes_per_measure) + end_of_group > notes_per_measure) {
-        // if we are in an odd time signature then the last few notes will have a different grouping to reach the end of the measure
-        end_of_group = notes_per_measure - (i % num_notes);
-      }
-
-      if (i % abc_gen_note_grouping_size(true, timeSigTop, timeSigBottom) === 0) {
-        // Look for some special cases that will format beats as non triplet groups.   Quarter notes, 1/8th and 1/16th notes only.
-
-        // look for a whole beat of rests
-        if (0 == count_active_notes_in_arrays(all_drum_array_of_array, i, 12)) {
-          // there are no notes in the next beat.   Let's output a special string for a quarter note rest
-          skip_adding_more_notes = true;
-          stickings_voice_string += 'x8';
-          hh_snare_voice_string += 'z8'; // quarter note rest
-          i += 11; // skip past all the rests
-
-          // look for 1/4 note with no triplets  "x--"
-        } else if (0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 1, 11)) {
-          // code duplicated from below
-          // clear any invalid stickings since they will mess up the formatting greatly
-          for (var si = i + 1; si < i + 12; si++) sticking_array[si] = false;
-          stickings_voice_string += getABCforRest([sticking_array], i, 8, scaler, true);
-          stickings_voice_string += getABCforNote([sticking_array], i, 8, scaler);
-
-          if (kick_stems_up) {
-            hh_snare_voice_string += getABCforNote(all_drum_array_of_array, i, 8, scaler);
-            kick_voice_string = '';
-          } else {
-            hh_snare_voice_string += getABCforNote(all_drum_array_of_array, i, 8, scaler);
-            kick_voice_string += getABCforNote([kick_array], i, 8, scaler);
-          }
-
-          skip_adding_more_notes = true;
-          i += 11; // skip past to the next beat
-
-          // look for two 1/8 notes with no triplets in 1/16th & 1/32nd note triplets.   "x--x--", "x-----x-----"
-        } else if (
-          sub_division > 12 &&
-          0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 1, 5) &&
-          0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 7, 5)
-        ) {
-          // think of the 1/8 notes as two groups of 3 notes
-          for (var eighth_index = i; eighth_index <= i + 6; eighth_index += 6) {
-            // code duplicated from below
-            // clear any invalid stickings since they will mess up the formatting greatly
-            for (si = eighth_index + 1; si < eighth_index + 6; si++) sticking_array[si] = false;
-            stickings_voice_string += getABCforRest(
-              [sticking_array],
-              eighth_index,
-              4,
-              scaler,
-              true
-            );
-            stickings_voice_string += getABCforNote([sticking_array], eighth_index, 4, scaler);
-
-            if (kick_stems_up) {
-              hh_snare_voice_string += getABCforRest(
-                all_drum_array_of_array,
-                eighth_index,
-                4,
-                scaler,
-                false
-              );
-              hh_snare_voice_string += getABCforNote(
-                all_drum_array_of_array,
-                eighth_index,
-                4,
-                scaler
-              );
-              kick_voice_string = '';
-            } else {
-              hh_snare_voice_string += getABCforRest(
-                all_drum_array_of_array,
-                eighth_index,
-                4,
-                scaler,
-                false
-              );
-              hh_snare_voice_string += getABCforNote(
-                all_drum_array_of_array,
-                eighth_index,
-                4,
-                scaler
-              );
-              kick_voice_string += getABCforNote([kick_array], eighth_index, 4, scaler);
-            }
-          }
-
-          skip_adding_more_notes = true;
-          i += 11; // skip past to the next beat
-
-          // look for 1/16th notes with no triplets in 1/32nd note triplets.   "x--x--"
-        } else if (
-          sub_division == 48 &&
-          0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 1, 2) &&
-          0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 4, 2) &&
-          0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 7, 2) &&
-          0 == count_active_notes_in_arrays(all_drum_array_of_array, i + 10, 2)
-        ) {
-          // think of the 1/8 notes as two groups of 3 notes
-          for (eighth_index = i; eighth_index <= i + 9; eighth_index += 3) {
-            // code duplicated from below
-            // clear any invalid stickings since they will mess up the formatting greatly
-            for (si = eighth_index + 1; si < eighth_index + 3; si++) sticking_array[si] = false;
-            stickings_voice_string += getABCforRest(
-              [sticking_array],
-              eighth_index,
-              2,
-              scaler,
-              true
-            );
-            stickings_voice_string += getABCforNote([sticking_array], eighth_index, 2, scaler);
-
-            if (kick_stems_up) {
-              hh_snare_voice_string += getABCforRest(
-                all_drum_array_of_array,
-                eighth_index,
-                2,
-                scaler,
-                false
-              );
-              hh_snare_voice_string += getABCforNote(
-                all_drum_array_of_array,
-                eighth_index,
-                2,
-                scaler
-              );
-              kick_voice_string = '';
-            } else {
-              hh_snare_voice_string += getABCforRest(
-                all_drum_array_of_array,
-                eighth_index,
-                2,
-                scaler,
-                false
-              );
-              hh_snare_voice_string += getABCforNote(
-                all_drum_array_of_array,
-                eighth_index,
-                2,
-                scaler
-              );
-              kick_voice_string += getABCforNote([kick_array], eighth_index, 2, scaler);
-            }
-          }
-
-          skip_adding_more_notes = true;
-          i += 11; // skip past to the next beat
-        } else {
-          // the normal case.   We tell ABC that we are using a triplet
-          var notes_in_triplet_group = sub_division / 4; // 4 beats
-
-          // look through the notes and see if we should "fake" 1/8 or 1/6th note triplets
-          // If the groove can be expressed in "3" or "6" groups it is way easier to read than in a higher "12" group with rests
-          // "3" looks like "x---x---x---"   one note and three rests
-          // "6" looks like "x-x-x-x-x-x-"   one note and one rest
-          if (sub_division == 48) {
-            var can_fake_threes = true;
-            var can_fake_sixes = true;
-            for (var j = i; j < i + 12; j += 4) {
-              if (0 < count_active_notes_in_arrays(all_drum_array_of_array, j + 1, 3)) {
-                can_fake_threes = false;
-              }
-              if (
-                0 < count_active_notes_in_arrays(all_drum_array_of_array, j + 1, 1) ||
-                0 < count_active_notes_in_arrays(all_drum_array_of_array, j + 3, 1)
-              ) {
-                can_fake_sixes = false;
-              }
-              if (can_fake_threes == false && can_fake_sixes == false) break; // skip the rest, since we have an answer already
-            }
-
-            // reset
-
-            if (can_fake_threes) faker_sub_division = 12;
-            else if (can_fake_sixes) faker_sub_division = 24;
-            else faker_sub_division = sub_division; // reset
-
-            end_of_group = 48 / faker_sub_division;
-            grouping_size_for_rests = end_of_group;
-            notes_in_triplet_group = faker_sub_division / 4; // 4 beats
-          }
-
-          // creates the 3, 6 or 12 over the note grouping
-          // looks like (3:3:3 or (6:6:6 or (12:12:12
-          hh_snare_voice_string +=
-            '(' +
-            notes_in_triplet_group +
-            ':' +
-            notes_in_triplet_group +
-            ':' +
-            notes_in_triplet_group;
-        }
-      }
-
-      // skip the code to add notes
-      // Happens for special_rest when there are no notes for the next whole beat.
-      // Happens when we found only a 1/4 or 1/8 note instead of triplets
-      if (!skip_adding_more_notes) {
-        if (i % grouping_size_for_rests === 0) {
-          // we will output a rest for each place there could be a note
-          stickings_voice_string += getABCforRest(
-            [sticking_array],
-            i,
-            grouping_size_for_rests,
-            scaler,
-            true
-          );
-
-          if (kick_stems_up) {
-            hh_snare_voice_string += getABCforRest(
-              all_drum_array_of_array,
-              i,
-              grouping_size_for_rests,
-              scaler,
-              false
-            );
-            kick_voice_string = '';
-          } else {
-            hh_snare_voice_string += getABCforRest(
-              all_drum_array_of_array,
-              i,
-              grouping_size_for_rests,
-              scaler,
-              false
-            );
-            kick_voice_string += getABCforRest(
-              [kick_array],
-              i,
-              grouping_size_for_rests,
-              scaler,
-              true
-            );
-          }
-        }
-
-        stickings_voice_string += getABCforNote([sticking_array], i, end_of_group, scaler);
-
-        if (kick_stems_up) {
-          hh_snare_voice_string += getABCforNote(all_drum_array_of_array, i, end_of_group, scaler);
-          kick_voice_string = '';
-        } else {
-          hh_snare_voice_string += getABCforNote(all_drum_array_of_array, i, end_of_group, scaler);
-          kick_voice_string += getABCforNote([kick_array], i, end_of_group, scaler);
-        }
-      }
-
-      if (
-        i % abc_gen_note_grouping_size(true, timeSigTop, timeSigBottom) ==
-        abc_gen_note_grouping_size(true, timeSigTop, timeSigBottom) - 1
-      ) {
-        stickings_voice_string += ' ';
-        hh_snare_voice_string += ' '; // Add a space to break the bar line every group notes
-        kick_voice_string += ' ';
-      }
-
-      // add a bar line every measure
-      if ((i + 1) % (12 * timeSigTop * (4 / timeSigBottom)) === 0) {
-        stickings_voice_string += '|';
-        hh_snare_voice_string += '|';
-        kick_voice_string += '|';
-
-        // add a line break every numberOfMeasuresPerLine measures
-        if (
-          i < num_notes - 1 &&
-          (i + 1) % (12 * timeSigTop * (4 / timeSigBottom) * numberOfMeasuresPerLine) === 0
-        ) {
-          stickings_voice_string += '\n';
-          hh_snare_voice_string += '\n';
-          kick_voice_string += '\n';
-        }
-      }
-    }
-
-    if (kick_stems_up)
-      ABC_String +=
-        stickings_voice_string + post_voice_abc + hh_snare_voice_string + post_voice_abc;
-    else
-      ABC_String +=
-        stickings_voice_string +
-        post_voice_abc +
-        hh_snare_voice_string +
-        post_voice_abc +
-        kick_voice_string +
-        post_voice_abc;
-
-    return ABC_String;
-  }
 
   // takes 4 arrays 32 elements long that represent the sticking, snare, HH & kick.
   // each element contains either the note value in ABC "F","^g" or false to represent off
   // translates them to an ABC string in 3 voices
   // post_voice_abc is a string added to the end of each voice line that can end the line
   //
-  function snare_HH_kick_ABC_for_quads(
-    sticking_array,
-    HH_array,
-    snare_array,
-    kick_array,
-    toms_array,
-    post_voice_abc,
-    num_notes,
-    sub_division,
-    notes_per_measure,
-    kick_stems_up,
-    timeSigTop,
-    timeSigBottom,
-    numberOfMeasuresPerLine
-  ) {
-    var scaler = 1; // we are always in 32ths notes here
-    var ABC_String = '';
-    var stickings_voice_string = 'V:Stickings\n'; // for stickings.  they are all rests with text comments added
-    var hh_snare_voice_string = 'V:Hands stem=up\n%%voicemap drum\n'; // for hh and snare
-    var kick_voice_string = 'V:Feet stem=down\n%%voicemap drum\n'; // for kick drum
-    var all_drum_array_of_array;
-
-    all_drum_array_of_array = [snare_array, HH_array]; // exclude the kick
-    if (toms_array) all_drum_array_of_array = all_drum_array_of_array.concat(toms_array);
-    // Add the kick array last to solve a subtle bug with the kick foot splash combo note
-    // If the combo note comes last in a multi note event it will space correctly.  If it
-    // comes first it will create a wrong sized note
-    if (kick_stems_up) all_drum_array_of_array = all_drum_array_of_array.concat([kick_array]);
-
-    for (var i = 0; i < num_notes; i++) {
-      var grouping_size_for_rests = abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom);
-      // make sure the group end doesn't go beyond the measure.   Happens in odd time sigs
-      if ((i % notes_per_measure) + grouping_size_for_rests > notes_per_measure) {
-        // if we are in an odd time signature then the last few notes will have a different grouping to reach the end of the measure
-        grouping_size_for_rests = notes_per_measure - (i % notes_per_measure);
-      }
-
-      var end_of_group;
-      if (i % abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom) === 0)
-        end_of_group = abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom);
-      else
-        end_of_group =
-          abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom) -
-          (i % abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom));
-
-      // make sure the group end doesn't go beyond the measure.   Happens in odd time sigs
-      if ((i % notes_per_measure) + end_of_group > notes_per_measure) {
-        // if we are in an odd time signature then the last few notes will have a different grouping to reach the end of the measure
-        end_of_group = notes_per_measure - (i % notes_per_measure);
-      }
-
-      if (i % abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom) === 0) {
-        // we will only output a rest at the beginning of a beat phrase
-        stickings_voice_string += getABCforRest(
-          [sticking_array],
-          i,
-          grouping_size_for_rests,
-          scaler,
-          true
-        );
-
-        if (kick_stems_up) {
-          hh_snare_voice_string += getABCforRest(
-            all_drum_array_of_array,
-            i,
-            grouping_size_for_rests,
-            scaler,
-            false
-          );
-          kick_voice_string = '';
-        } else {
-          hh_snare_voice_string += getABCforRest(
-            all_drum_array_of_array,
-            i,
-            grouping_size_for_rests,
-            scaler,
-            false
-          );
-          kick_voice_string += getABCforRest(
-            [kick_array],
-            i,
-            grouping_size_for_rests,
-            scaler,
-            false
-          );
-        }
-      }
-
-      stickings_voice_string += getABCforNote([sticking_array], i, end_of_group, scaler);
-
-      if (kick_stems_up) {
-        hh_snare_voice_string += getABCforNote(all_drum_array_of_array, i, end_of_group, scaler);
-        kick_voice_string = '';
-      } else {
-        hh_snare_voice_string += getABCforNote(all_drum_array_of_array, i, end_of_group, scaler);
-        kick_voice_string += getABCforNote([kick_array], i, end_of_group, scaler);
-      }
-
-      if (
-        i % abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom) ==
-        abc_gen_note_grouping_size(false, timeSigTop, timeSigBottom) - 1
-      ) {
-        stickings_voice_string += ' ';
-        hh_snare_voice_string += ' '; // Add a space to break the bar line every group notes
-        kick_voice_string += ' ';
-      }
-
-      // add a bar line every measure.   32 notes in 4/4 time.   (32/timeSigBottom * timeSigTop)
-      if ((i + 1) % ((32 / timeSigBottom) * timeSigTop) === 0) {
-        stickings_voice_string += '|';
-        hh_snare_voice_string += '|';
-        kick_voice_string += '|';
-      }
-      // add a line break every numberOfMeasuresPerLine measures, except the last
-      if (
-        i < num_notes - 1 &&
-        (i + 1) % ((32 / timeSigBottom) * timeSigTop * numberOfMeasuresPerLine) === 0
-      ) {
-        stickings_voice_string += '\n';
-        hh_snare_voice_string += '\n';
-        kick_voice_string += '\n';
-      }
-    }
-
-    if (kick_stems_up)
-      ABC_String +=
-        stickings_voice_string + post_voice_abc + hh_snare_voice_string + post_voice_abc;
-    else
-      ABC_String +=
-        stickings_voice_string +
-        post_voice_abc +
-        hh_snare_voice_string +
-        post_voice_abc +
-        kick_voice_string +
-        post_voice_abc;
-
-    return ABC_String;
-  }
 
   // create an array that can be used for note mapping
   // it is just an array of true/false that specifies weather a note can appear at that index
@@ -2308,149 +1329,28 @@ function GrooveUtils() {
     timeSigTop,
     timeSigBottom
   ) {
-    // convert sticking count symbol to the actual count
-    // do this right before ABC output so it can't every get encoded into something that gets saved.
-    root.convert_sticking_counts_to_actual_counts(
+    return _create_ABC_from_snare_HH_kick_arrays(
+      root,
       sticking_array,
+      HH_array,
+      snare_array,
+      kick_array,
+      toms_array,
+      post_voice_abc,
+      num_notes,
       time_division,
+      notes_per_measure,
+      kick_stems_up,
       timeSigTop,
       timeSigBottom
     );
-
-    var numberOfMeasuresPerLine = 2; // Default
-
-    if (notes_per_measure >= 32) {
-      // Only put one measure per line for 32nd notes and above because of width issues
-      numberOfMeasuresPerLine = 1;
-    }
-
-    if (root.isTripletDivisionFromNotesPerMeasure(notes_per_measure, timeSigTop, timeSigBottom)) {
-      return snare_HH_kick_ABC_for_triplets(
-        sticking_array,
-        HH_array,
-        snare_array,
-        kick_array,
-        toms_array,
-        post_voice_abc,
-        num_notes,
-        time_division,
-        notes_per_measure,
-        kick_stems_up,
-        timeSigTop,
-        timeSigBottom,
-        numberOfMeasuresPerLine
-      );
-    } else {
-      return snare_HH_kick_ABC_for_quads(
-        sticking_array,
-        HH_array,
-        snare_array,
-        kick_array,
-        toms_array,
-        post_voice_abc,
-        num_notes,
-        time_division,
-        notes_per_measure,
-        kick_stems_up,
-        timeSigTop,
-        timeSigBottom,
-        numberOfMeasuresPerLine
-      );
-    }
   };
 
   // create ABC notation from a GrooveData class
   // returns a string of ABC Notation data
 
   root.createABCFromGrooveData = function (myGrooveData, renderWidth) {
-    var FullNoteStickingArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.sticking_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-    var FullNoteHHArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.hh_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-    var FullNoteSnareArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.snare_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-    var FullNoteKickArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.kick_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-    var FullNoteTomsArray = [];
-
-    for (var i = 0; i < constant_NUMBER_OF_TOMS; i++) {
-      FullNoteTomsArray[i] = root.scaleNoteArrayToFullSize(
-        myGrooveData.toms_array[i],
-        myGrooveData.numberOfMeasures,
-        myGrooveData.notesPerMeasure,
-        myGrooveData.numBeats,
-        myGrooveData.noteValue
-      );
-    }
-
-    var is_triplet_division = root.isTripletDivisionFromNotesPerMeasure(
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-
-    var fullABC = root.get_top_ABC_BoilerPlate(
-      false,
-      myGrooveData.title,
-      myGrooveData.author,
-      myGrooveData.comments,
-      myGrooveData.showLegend,
-      is_triplet_division,
-      myGrooveData.kickStemsUp,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue,
-      renderWidth
-    );
-
-    fullABC += root.create_ABC_from_snare_HH_kick_arrays(
-      FullNoteStickingArray,
-      FullNoteHHArray,
-      FullNoteSnareArray,
-      FullNoteKickArray,
-      FullNoteTomsArray,
-      '|\n',
-      FullNoteHHArray.length,
-      myGrooveData.timeDivision,
-      root.notesPerMeasureInFullSizeArray(
-        is_triplet_division,
-        myGrooveData.numBeats,
-        myGrooveData.noteValue
-      ), // notes_per_measure, We scaled up to 48/32 above
-      myGrooveData.kickStemsUp,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-
-    root.note_mapping_array = root.create_note_mapping_array_for_highlighting(
-      FullNoteHHArray,
-      FullNoteSnareArray,
-      FullNoteKickArray,
-      FullNoteTomsArray,
-      FullNoteHHArray.length
-    );
-
-    // console.log(fullABC);
-    return fullABC;
+    return _createABCFromGrooveData(root, myGrooveData, renderWidth);
   };
 
   // callback class for abc generator library
@@ -2720,39 +1620,7 @@ function GrooveUtils() {
   };
 
   root.MIDI_build_midi_url_count_in_track = function (timeSigTop, timeSigBottom) {
-    var midiFile = new Midi.File();
-    var midiTrack = new Midi.Track();
-    midiFile.addTrack(midiTrack);
-
-    midiTrack.setTempo(root.getTempo());
-    midiTrack.setInstrument(0, 0x13);
-
-    // start of midi track
-    // Some sort of bug in the midi player makes it skip the first note without a blank
-    // TODO: Find and fix midi bug
-    midiTrack.addNoteOff(9, 60, 1); // add a blank note for spacing
-
-    var noteDelay = 128; // quarter notes over x/4 time
-    if (timeSigBottom == 8)
-      noteDelay = 64; // 8th notes over x/8 time
-    else if (timeSigBottom == 16) noteDelay = 32; // 16th notes over x/16 time
-
-    // add count in
-    midiTrack.addNoteOn(9, constant_OUR_MIDI_METRONOME_1, 0, constant_OUR_MIDI_VELOCITY_NORMAL);
-    midiTrack.addNoteOff(9, constant_OUR_MIDI_METRONOME_1, noteDelay);
-    for (var i = 1; i < timeSigTop; i++) {
-      midiTrack.addNoteOn(
-        9,
-        constant_OUR_MIDI_METRONOME_NORMAL,
-        0,
-        constant_OUR_MIDI_VELOCITY_NORMAL
-      );
-      midiTrack.addNoteOff(9, constant_OUR_MIDI_METRONOME_NORMAL, noteDelay);
-    }
-
-    var midi_url = 'data:audio/midi;base64,' + btoa(midiFile.toBytes());
-
-    return midi_url;
+    return _MIDI_build_midi_url_count_in_track(root, timeSigTop, timeSigBottom);
   };
 
   /*
@@ -2779,438 +1647,26 @@ function GrooveUtils() {
     timeSigTop,
     timeSigBottom
   ) {
-    var prev_hh_note = 46; // default to open hi-hat so that the first hi-hat note also mutes any previous hh open.
-    var midi_channel = 9; // percussion
-
-    if (swing_percentage < 0 || swing_percentage > 0.99) {
-      console.log('Swing percentage out of range in GrooveUtils.MIDI_from_HH_Snare_Kick_Arrays');
-      swing_percentage = 0;
-    }
-
-    // start of midi track
-    // Some sort of bug in the midi player makes it skip the first note without a blank
-    // TODO: Find and fix midi bug
-    if (midiTrack.events.length < 4) {
-      midiTrack.addNoteOff(midi_channel, 60, 1); // add a blank note for spacing
-    }
-
-    var isTriplets = root.isTripletDivisionFromNotesPerMeasure(
+    return _MIDI_from_HH_Snare_Kick_Arrays(
+      root,
+      midiTrack,
+      HH_Array,
+      Snare_Array,
+      Kick_Array,
+      Toms_Array,
+      midi_output_type,
+      metronome_frequency,
       num_notes,
+      num_notes_for_swing,
+      swing_percentage,
       timeSigTop,
       timeSigBottom
     );
-    var offsetClickStartBeat = root.getMetronomeOptionsOffsetClickStartRotation(isTriplets);
-    var delay_for_next_note = 0;
-
-    for (var i = 0; i < num_notes; i++) {
-      var duration = 0;
-
-      if (isTriplets) {
-        duration = 10.666; // "ticks"   16 for 32nd notes.  10.66 for 48th triplets
-      } else {
-        duration = 16;
-      }
-
-      if (swing_percentage !== 0) {
-        // swing effects the note placement of the e and the a.  (1e&a)
-        // swing increases the distance between the 1 and the e ad shortens the distance between the e and the &
-        // likewise the distance between the & and the a is increased and the a and the 1 is shortened
-        //  So it sounds like this:   1-e&-a2-e&-a3-e&-a4-e&-a
-        var scaler = num_notes / num_notes_for_swing;
-        var val = i % (4 * scaler);
-
-        if (val < scaler) {
-          // this is the 1, increase the distance between this note and the e
-          duration += duration * swing_percentage;
-        } else if (val < scaler * 2) {
-          // this is the e, shorten the distance between this note and the &
-          duration -= duration * swing_percentage;
-        } else if (val < scaler * 3) {
-          // this is the &, increase the distance between this note and the a
-          duration += duration * swing_percentage;
-        } else if (val < scaler * 4) {
-          // this is the a, shorten the distance between this note and the 2
-          duration -= duration * swing_percentage;
-        }
-      }
-
-      // Metronome sounds.
-      var metronome_note = false;
-      var metronome_velocity = constant_OUR_MIDI_VELOCITY_ACCENT;
-      if (metronome_frequency > 0) {
-        var quarterNoteFrequency = isTriplets ? 12 : 8;
-        var eighthNoteFrequency = isTriplets ? 6 : 4;
-        var sixteenthNoteFrequency = isTriplets ? 2 : 2;
-
-        var metronome_specific_index = i;
-        switch (offsetClickStartBeat) {
-          case '1':
-            // default do nothing
-            break;
-          case 'E':
-            if (isTriplets) console.log('OffsetClickStart error in MIDI_from_HH_Snare_Kick_Arrays');
-            // shift by one sixteenth note
-            metronome_specific_index -= sixteenthNoteFrequency;
-            break;
-          case 'AND':
-            if (isTriplets) console.log('OffsetClickStart error in MIDI_from_HH_Snare_Kick_Arrays');
-            // shift by two sixteenth notes
-            metronome_specific_index -= 2 * sixteenthNoteFrequency;
-            break;
-          case 'A':
-            if (isTriplets) console.log('OffsetClickStart error in MIDI_from_HH_Snare_Kick_Arrays');
-            // shift by three sixteenth notes
-            metronome_specific_index -= 3 * sixteenthNoteFrequency;
-            break;
-          case 'TI':
-            if (!isTriplets)
-              console.log('OffsetClickStart error in MIDI_from_HH_Snare_Kick_Arrays');
-            // shift by one sixteenth note
-            metronome_specific_index -= sixteenthNoteFrequency * 2;
-            break;
-          case 'TA':
-            if (!isTriplets)
-              console.log('OffsetClickStart error in MIDI_from_HH_Snare_Kick_Arrays');
-            // shift by two sixteenth notes
-            metronome_specific_index -= 2 * (sixteenthNoteFrequency * 2);
-            break;
-          default:
-            console.log('bad case in MIDI_from_HH_Snare_Kick_Arrays');
-            break;
-        }
-
-        if (metronome_specific_index >= 0) {
-          // can go negative due to MetronomeOffsetClickStart shift above
-          // Special sound on the one
-          if (
-            metronome_specific_index === 0 ||
-            metronome_specific_index % (quarterNoteFrequency * timeSigTop * (4 / timeSigBottom)) ===
-              0
-          ) {
-            metronome_note = constant_OUR_MIDI_METRONOME_1; // 1 count
-          } else if (metronome_specific_index % quarterNoteFrequency === 0) {
-            metronome_note = constant_OUR_MIDI_METRONOME_NORMAL; // standard metronome click
-          }
-
-          if (!metronome_note && metronome_frequency == 8) {
-            // 8th notes requested
-            if (metronome_specific_index % eighthNoteFrequency === 0) {
-              // click every 8th note
-              metronome_note = constant_OUR_MIDI_METRONOME_NORMAL; // standard metronome click
-            }
-          } else if (!metronome_note && metronome_frequency == 16) {
-            // 16th notes requested
-            if (metronome_specific_index % sixteenthNoteFrequency === 0) {
-              // click every 16th note
-              metronome_note = constant_OUR_MIDI_METRONOME_NORMAL; // standard metronome click
-              metronome_velocity = 25; // not as loud as the normal click
-            }
-          }
-        }
-
-        if (metronome_note !== false) {
-          //if(prev_metronome_note != false)
-          //	midiTrack.addNoteOff(midi_channel, prev_metronome_note, 0);
-          midiTrack.addNoteOn(
-            midi_channel,
-            metronome_note,
-            delay_for_next_note,
-            metronome_velocity
-          );
-          delay_for_next_note = 0; // zero the delay
-          //prev_metronome_note = metronome_note;
-        }
-      }
-
-      if (!root.metronomeSolo) {
-        // midiSolo means to play just the metronome
-        var hh_velocity = constant_OUR_MIDI_VELOCITY_NORMAL;
-        var hh_note = false;
-        switch (HH_Array[i]) {
-          case constant_ABC_HH_Normal: // normal
-          case constant_ABC_HH_Close: // normal
-            hh_note = constant_OUR_MIDI_HIHAT_NORMAL;
-            break;
-          case constant_ABC_HH_Accent: // accent
-            if (midi_output_type == 'general_MIDI') {
-              hh_note = constant_OUR_MIDI_HIHAT_NORMAL;
-              hh_velocity = constant_OUR_MIDI_VELOCITY_ACCENT;
-            } else {
-              hh_note = constant_OUR_MIDI_HIHAT_ACCENT;
-            }
-            break;
-          case constant_ABC_HH_Open: // open
-            hh_note = constant_OUR_MIDI_HIHAT_OPEN;
-            break;
-          case constant_ABC_HH_Ride: // ride
-            hh_note = constant_OUR_MIDI_HIHAT_RIDE;
-            break;
-          case constant_ABC_HH_Ride_Bell: // ride bell
-            hh_note = constant_OUR_MIDI_HIHAT_RIDE_BELL;
-            break;
-          case constant_ABC_HH_Cow_Bell: // cow bell
-            hh_note = constant_OUR_MIDI_HIHAT_COW_BELL;
-            break;
-          case constant_ABC_HH_Crash: // crash
-            hh_note = constant_OUR_MIDI_HIHAT_CRASH;
-            break;
-          case constant_ABC_HH_Stacker: // stacker
-            hh_note = constant_OUR_MIDI_HIHAT_STACKER;
-            break;
-          case constant_ABC_HH_Metronome_Normal: // Metronome beep
-            hh_note = constant_OUR_MIDI_HIHAT_METRONOME_NORMAL;
-            break;
-          case constant_ABC_HH_Metronome_Accent: // Metronome beep
-            hh_note = constant_OUR_MIDI_HIHAT_METRONOME_ACCENT;
-            break;
-          case false:
-            break;
-          default:
-            console.log('Bad case in GrooveUtils.MIDI_from_HH_Snare_Kick_Arrays');
-            break;
-        }
-
-        if (hh_note !== false) {
-          // need to end hi-hat open notes else the hh open sounds horrible
-          if (prev_hh_note !== false) {
-            midiTrack.addNoteOff(midi_channel, prev_hh_note, delay_for_next_note);
-            prev_hh_note = false;
-            delay_for_next_note = 0; // zero the delay
-          }
-          midiTrack.addNoteOn(midi_channel, hh_note, delay_for_next_note, hh_velocity);
-          delay_for_next_note = 0; // zero the delay
-
-          // this if means that only the open hi-hat will get stopped on the next note
-          if (HH_Array[i] == constant_ABC_HH_Open) prev_hh_note = hh_note;
-        }
-
-        var snare_velocity = constant_OUR_MIDI_VELOCITY_NORMAL;
-        var snare_note = false;
-        switch (Snare_Array[i]) {
-          case constant_ABC_SN_Normal: // normal
-            snare_note = constant_OUR_MIDI_SNARE_NORMAL;
-            break;
-          case constant_ABC_SN_Flam: // flam
-            if (midi_output_type == 'general_MIDI') {
-              snare_note = constant_OUR_MIDI_SNARE_NORMAL;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_ACCENT;
-            } else {
-              snare_note = constant_OUR_MIDI_SNARE_FLAM;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_NORMAL;
-            }
-            break;
-          case constant_ABC_SN_Drag: // drag
-            if (midi_output_type == 'general_MIDI') {
-              snare_note = constant_OUR_MIDI_SNARE_NORMAL;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_ACCENT;
-            } else {
-              snare_note = constant_OUR_MIDI_SNARE_DRAG;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_NORMAL;
-            }
-            break;
-          case constant_ABC_SN_Accent: // accent
-            if (midi_output_type == 'general_MIDI') {
-              snare_note = constant_OUR_MIDI_SNARE_NORMAL;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_ACCENT;
-            } else {
-              snare_note = constant_OUR_MIDI_SNARE_ACCENT; // custom note
-            }
-            break;
-          case constant_ABC_SN_Ghost: // ghost
-            if (midi_output_type == 'general_MIDI') {
-              snare_note = constant_OUR_MIDI_SNARE_NORMAL;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_GHOST;
-            } else {
-              snare_note = constant_OUR_MIDI_SNARE_GHOST;
-              snare_velocity = constant_OUR_MIDI_VELOCITY_GHOST;
-            }
-            break;
-          case constant_ABC_SN_XStick: // xstick
-            snare_note = constant_OUR_MIDI_SNARE_XSTICK;
-            break;
-          case constant_ABC_SN_Buzz: // xstick
-            snare_note = constant_OUR_MIDI_SNARE_BUZZ;
-            break;
-          case false:
-            break;
-          default:
-            console.log('Bad case in GrooveUtils.MIDI_from_HH_Snare_Kick_Arrays');
-            break;
-        }
-
-        if (snare_note !== false) {
-          //if(prev_snare_note != false)
-          //	midiTrack.addNoteOff(midi_channel, prev_snare_note, 0);
-          midiTrack.addNoteOn(midi_channel, snare_note, delay_for_next_note, snare_velocity);
-          delay_for_next_note = 0; // zero the delay
-          //prev_snare_note = snare_note;
-        }
-
-        var kick_note = false;
-        var kick_splash_note = false;
-        switch (Kick_Array[i]) {
-          case constant_ABC_KI_Splash: // just HH Foot
-            kick_splash_note = constant_OUR_MIDI_HIHAT_FOOT;
-            break;
-          case constant_ABC_KI_SandK: // Kick & HH Foot
-            kick_splash_note = constant_OUR_MIDI_HIHAT_FOOT;
-            kick_note = constant_OUR_MIDI_KICK_NORMAL;
-            break;
-          case constant_ABC_KI_Normal: // just Kick
-            kick_note = constant_OUR_MIDI_KICK_NORMAL;
-            break;
-          case false:
-            break;
-          default:
-            console.log('Bad case in GrooveUtils.MIDI_from_HH_Snare_Kick_Arrays');
-            break;
-        }
-        if (kick_note !== false) {
-          //if(prev_kick_note != false)
-          //	midiTrack.addNoteOff(midi_channel, prev_kick_note, 0);
-          midiTrack.addNoteOn(
-            midi_channel,
-            kick_note,
-            delay_for_next_note,
-            constant_OUR_MIDI_VELOCITY_NORMAL
-          );
-          delay_for_next_note = 0; // zero the delay
-          //prev_kick_note = kick_note;
-        }
-        if (kick_splash_note !== false) {
-          if (prev_hh_note !== false) {
-            midiTrack.addNoteOff(midi_channel, prev_hh_note, delay_for_next_note);
-            prev_hh_note = false;
-            delay_for_next_note = 0; // zero the delay
-          }
-          //if(prev_kick_splash_note != false)
-          //	midiTrack.addNoteOff(midi_channel, prev_kick_splash_note, 0);
-          midiTrack.addNoteOn(
-            midi_channel,
-            kick_splash_note,
-            delay_for_next_note,
-            constant_OUR_MIDI_VELOCITY_NORMAL
-          );
-          delay_for_next_note = 0; // zero the delay
-          //prev_kick_splash_note = kick_splash_note;
-        }
-
-        if (Toms_Array) {
-          for (var which_array = 0; which_array < constant_NUMBER_OF_TOMS; which_array++) {
-            var tom_note = false;
-            if (Toms_Array[which_array][i] !== undefined) {
-              switch (Toms_Array[which_array][i]) {
-                case constant_ABC_T1_Normal: // Tom 1
-                  tom_note = constant_OUR_MIDI_TOM1_NORMAL; // midi code High tom 2
-                  break;
-                case constant_ABC_T2_Normal: // Midi code Mid tom 1
-                  tom_note = constant_OUR_MIDI_TOM2_NORMAL;
-                  break;
-                case constant_ABC_T3_Normal: // Midi code Mid tom 2
-                  tom_note = constant_OUR_MIDI_TOM3_NORMAL;
-                  break;
-                case constant_ABC_T4_Normal: // Midi code Low Tom 1
-                  tom_note = constant_OUR_MIDI_TOM4_NORMAL;
-                  break;
-                case false:
-                  break;
-                default:
-                  console.log('Bad case in GrooveUtils.MIDI_from_HH_Snare_Kick_Arrays');
-                  break;
-              }
-            }
-            if (tom_note !== false) {
-              midiTrack.addNoteOn(
-                midi_channel,
-                tom_note,
-                delay_for_next_note,
-                constant_OUR_MIDI_VELOCITY_NORMAL
-              );
-              delay_for_next_note = 0; // zero the delay
-            }
-          }
-        }
-      } // end metronomeSolo
-
-      delay_for_next_note += duration;
-    }
-
-    if (delay_for_next_note) midiTrack.addNoteOff(0, 60, delay_for_next_note - 1); // add a blank note for spacing
-  }; // end of function
+  };
 
   // returns a URL that is a MIDI track
   root.create_MIDIURLFromGrooveData = function (myGrooveData, MIDI_type) {
-    var midiFile = new Midi.File();
-    var midiTrack = new Midi.Track();
-    midiFile.addTrack(midiTrack);
-
-    midiTrack.setTempo(myGrooveData.tempo);
-    midiTrack.setInstrument(0, 0x13);
-
-    var swing_percentage = myGrooveData.swingPercent / 100;
-
-    // the midi converter expects all the arrays to be 32 or 48 notes long.
-    // Expand them
-    var FullNoteHHArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.hh_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-    var FullNoteSnareArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.snare_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-    var FullNoteKickArray = root.scaleNoteArrayToFullSize(
-      myGrooveData.kick_array,
-      myGrooveData.numberOfMeasures,
-      myGrooveData.notesPerMeasure,
-      myGrooveData.numBeats,
-      myGrooveData.noteValue
-    );
-
-    // the midi functions expect just one measure at a time to work correctly
-    // call once for each measure
-    var measure_notes = FullNoteHHArray.length / myGrooveData.numberOfMeasures;
-    for (var measureIndex = 0; measureIndex < myGrooveData.numberOfMeasures; measureIndex++) {
-      var FullNoteTomsArray = [];
-      for (var i = 0; i < constant_NUMBER_OF_TOMS; i++) {
-        var orig_measure_notes = myGrooveData.notesPerMeasure;
-        FullNoteTomsArray[i] = root.scaleNoteArrayToFullSize(
-          myGrooveData.toms_array[i].slice(
-            orig_measure_notes * measureIndex,
-            orig_measure_notes * (measureIndex + 1)
-          ),
-          1,
-          myGrooveData.notesPerMeasure,
-          myGrooveData.numBeats,
-          myGrooveData.noteValue
-        );
-      }
-
-      root.MIDI_from_HH_Snare_Kick_Arrays(
-        midiTrack,
-        FullNoteHHArray.slice(measure_notes * measureIndex, measure_notes * (measureIndex + 1)),
-        FullNoteSnareArray.slice(measure_notes * measureIndex, measure_notes * (measureIndex + 1)),
-        FullNoteKickArray.slice(measure_notes * measureIndex, measure_notes * (measureIndex + 1)),
-        FullNoteTomsArray,
-        MIDI_type,
-        myGrooveData.metronomeFrequency,
-        measure_notes,
-        myGrooveData.timeDivision,
-        swing_percentage,
-        myGrooveData.numBeats,
-        myGrooveData.noteValue
-      );
-    }
-
-    var midi_url = 'data:audio/midi;base64,' + btoa(midiFile.toBytes());
-
-    return midi_url;
+    return _create_MIDIURLFromGrooveData(root, myGrooveData, MIDI_type);
   };
 
   root.loadMIDIFromURL = function (midiURL) {
@@ -3951,56 +2407,4 @@ function GrooveUtils() {
 // ES module exports: the GrooveUtils constructor plus the constants that other
 // modules (groove_writer) consume. Vendored globals (Abc, MIDI, Midi) remain
 // window globals provided by the classic <script> libraries.
-export {
-  GrooveUtils,
-  constant_ABC_HH_Accent,
-  constant_ABC_HH_Close,
-  constant_ABC_HH_Cow_Bell,
-  constant_ABC_HH_Crash,
-  constant_ABC_HH_Metronome_Accent,
-  constant_ABC_HH_Metronome_Normal,
-  constant_ABC_HH_Normal,
-  constant_ABC_HH_Open,
-  constant_ABC_HH_Ride,
-  constant_ABC_HH_Ride_Bell,
-  constant_ABC_HH_Stacker,
-  constant_ABC_KI_Normal,
-  constant_ABC_KI_SandK,
-  constant_ABC_KI_Splash,
-  constant_ABC_SN_Accent,
-  constant_ABC_SN_Buzz,
-  constant_ABC_SN_Drag,
-  constant_ABC_SN_Flam,
-  constant_ABC_SN_Ghost,
-  constant_ABC_SN_Normal,
-  constant_ABC_SN_XStick,
-  constant_ABC_STICK_BOTH,
-  constant_ABC_STICK_COUNT,
-  constant_ABC_STICK_L,
-  constant_ABC_STICK_OFF,
-  constant_ABC_STICK_R,
-  constant_ABC_T1_Normal,
-  constant_ABC_T4_Normal,
-  constant_OUR_MIDI_HIHAT_ACCENT,
-  constant_OUR_MIDI_HIHAT_COW_BELL,
-  constant_OUR_MIDI_HIHAT_CRASH,
-  constant_OUR_MIDI_HIHAT_FOOT,
-  constant_OUR_MIDI_HIHAT_METRONOME_ACCENT,
-  constant_OUR_MIDI_HIHAT_METRONOME_NORMAL,
-  constant_OUR_MIDI_HIHAT_NORMAL,
-  constant_OUR_MIDI_HIHAT_OPEN,
-  constant_OUR_MIDI_HIHAT_RIDE,
-  constant_OUR_MIDI_HIHAT_RIDE_BELL,
-  constant_OUR_MIDI_HIHAT_STACKER,
-  constant_OUR_MIDI_KICK_NORMAL,
-  constant_OUR_MIDI_SNARE_ACCENT,
-  constant_OUR_MIDI_SNARE_BUZZ,
-  constant_OUR_MIDI_SNARE_DRAG,
-  constant_OUR_MIDI_SNARE_FLAM,
-  constant_OUR_MIDI_SNARE_GHOST,
-  constant_OUR_MIDI_SNARE_NORMAL,
-  constant_OUR_MIDI_SNARE_XSTICK,
-  constant_OUR_MIDI_TOM1_NORMAL,
-  constant_OUR_MIDI_TOM4_NORMAL,
-  constant_OUR_MIDI_VELOCITY_NORMAL,
-};
+export { GrooveUtils };
